chore: further refactor
