--- src/chips/mod.rs
@@ -1,3 +1,0 @@
-pub mod poly_assigned;
-pub mod poly_distribution;
-pub mod utils;
